feat: discount voucher, use in purchases | fix: image dissapear when update item

+++ app/Http/Controllers/ItemController.php
@@ -77,14 +77,17 @@ class ItemController extends Controller
     {
         $data = $request->validated();
         $oldPath = "";
+        
         if ($request->hasFile("image")) {
             $file = $request->file("image")->store("images", "public");
             $oldPath = $item->image ?? "";
             $data["image"] = $file;
         }
-
+        
         if ($item->update($data)) {
-            Storage::disk('public')->delete($oldPath);
+            if ($request->hasFile("image")) {
+                Storage::disk('public')->delete($oldPath);
+            }
             return to_route("items.index")->with("success", "Success update item $item->name");
         }
